Refactor: move prompt caching logic to Anthropic handler and remove from base handler
